+++ HunterPie.Integrations/Datasources/MonsterHunterWorld/Definitions/MHWMapMonsterSelectionStructure.cs
@@ -5,7 +5,7 @@ namespace HunterPie.Integrations.Datasources.MonsterHunterWorld.Definitions;
 [StructLayout(LayoutKind.Explicit)]
 public struct MHWMapMonsterSelectionStructure
 {
-    [FieldOffset(0x128)] public long MapInsectsRef;
-    [FieldOffset(0x148)] public long SelectedMonster;
-    [FieldOffset(0x160)] public long GuiRadarRef;
+    [FieldOffset(0x128)] public nint MapInsectsRef;
+    [FieldOffset(0x148)] public nint SelectedMonster;
+    [FieldOffset(0x160)] public nint GuiRadarRef;
 }
